+++ src/afterglow/effects/params.clj
@@ -912,18 +912,4 @@
         (resolve-non-frame-dynamic-elements [this show snapshot head]
           (resolve-fn show snapshot head))))))
 
-(defn build-oscillated-param
-  "This function was moved
-  to [[afterglow.effects.oscillators/build-oscillated-param]] in
-  verson 0.1.6 in order to resolve circular dependency issues
-  introduced when Oscillators gained the ability to accept dynamic
-  parameters. This stub was left behind for backwards compatibility,
-  but will be removed in the next version, so please update your code
-  to find it in its new location."
-  {:deprecated "0.1.6"}
-  [osc & {:keys [min max metronome frame-dynamic] :or {min 0 max 255 frame-dynamic true}}]
-  (require '[afterglow.effects.oscillators])
-  ((resolve 'afterglow.effects.oscillators/build-oscillated-param) osc :min min :max max
-   :metronome metronome :frame-dynamic frame-dynamic))
-
 ;; TODO: some kind of random parameter?